Cinema.java's main method has been changed
By now, main method prints an seating arrangment depending
of the number of rows and seats from the input.
Then it read a row number and a seat number in that row
that represent the coordinates of the seat and prints
the ticket price according the coordinates

In Calculator.java created a new method that
calculate cost for seat in the seating arrangement.