Added a menu in a program
When you start the program
first, as before you enter the number of rows and seats
of the seats arrangement.
Then it shows a menu that allows to buy tickets and displays the
current state of the seating arrangement when needed.

Some methods for printng some information have been moved to
a new file called Output.java